requirements.txt
git+https://github.com/parelite/discord.py
git+https://github.com/Gorialis/jishaku
git+https://github.com/cloudwithax/pomice
pydub
PyNaCl
typing_extensions
humanize
rich
asyncpg
orjson
pydantic
humanfriendly
lru_dict
async_lru
redis
xxhash
beautifulsoup4
shazamio
python-dateutil
parsedatetime
aiohttp_cors
roblox
playwright
aiomisc
cryptography
msgpack
boltons
aiofile
tornado
psutil
wand
colorthief
httpx
asyncspotify
cashews
yt-dlp
pycryptodome
stackprinter
dateparser
arrow
xmltodict
cairosvg
asyncpraw
validators
rapidfuzz
lxml
colorama
aiomisc
git+https://github.com/MiroslavRosenov/better-ipc
pyparsing
pylast
wavelink